Fix zephyr build

+++ src/core/client/cache.c
@@ -28,7 +28,7 @@ in3_error_t in3_cache_update_nodelist(in3_t* c, in3_chain_t* chain) {
 
   // define the key to use
   char key[200];
-  sprintf(key, NODE_LIST_KEY, (unsigned long long) chain->chainId);
+  sprintf(key, NODE_LIST_KEY, chain->chainId);
 
   // get from cache
   bytes_t* b = c->cacheStorage->get_item(c->cacheStorage->cptr, key);
@@ -92,7 +92,7 @@ in3_error_t in3_cache_store_nodelist(in3_ctx_t* ctx, in3_chain_t* chain) {
 
   // create key
   char key[200];
-  sprintf(key, NODE_LIST_KEY, (unsigned long long) chain->chainId);
+  sprintf(key, NODE_LIST_KEY, chain->chainId);
 
   // store it and ignore return value since failing when writing cache should not stop us.
   ctx->client->cacheStorage->set_item(ctx->client->cacheStorage->cptr, key, &bb->b);
